Move SPMD compiler APIs to a separate folder.
