+++ spmd/__init__.py
@@ -5,7 +5,7 @@ from spmd.tensor import distribute_tensor
 from spmd.tensor.api import DTensor
 from spmd.tensor.device_mesh import get_global_device_mesh, DeviceMesh
 from spmd.tensor.placement_types import Replicate, Shard
-from spmd.api import Schema, SPMD
+from spmd.compiler.api import Schema, SPMD
 
 
 def distribute_module(
