+++ test/spmd/test_tracing.py
@@ -9,7 +9,7 @@ from torch.fx.experimental.proxy_tensor import make_fx
 from torch.nn.parallel import DistributedDataParallel as DDP
 from torch.testing._internal.common_utils import run_tests
 
-from spmd.api import SPMD, Schema
+from spmd.compiler.api import SPMD, Schema
 from spmd.testing.common_utils import (  # type: ignore
     DistTensorTestBase,
     with_comms,
